Remove CallSuper antipattern from RenderPreview::setupSceneGraph()
setupSceneGraph() is now an empty virtual method which subclasses are
free to override or not as they wish (all of them do, except
MapPreview).

The single line in the original default implementation (setting the
current render time to 0) is now moved up into the calling method, so
that subclasses are no longer required to call the parent implementation
to avoid breaking the base class behaviour.