Fix missing renderable boxes for certain entity types
Commit f99d1d5b417f043ee91b18 changed the return value of localAABB from
a const reference to a value. This in itself is not a problem, but the
RenderableBox class was expecting to be initialised with a reference to
a longer-lived AABB which it would render on demand. If the value
returned from localAABB was bound to this reference parameter, the code
would compile without any warnings, but the connection between the
source entity's bounding box and the RenderableBox was lost. This
resulted in entities rendering with zero or undefined bounding boxes.

The AABB parameter of the RenderableBox constructor is now a pointer,
which (at least on Linux) will not even compile if the calling code
attempts to bind it to a temporary.