TextureManipulator preferences now visible in dialog
TextureManipulator was a static object constructed on demand, and the
only call to construct it came from MapExpression when handling a
map expression that requires a texture to be resized. This meant that
under many circumstances the TextureManipulator was never constructed
and its constructPreferences method never called.

The TextureManipulator is now a component of MaterialManager and follows
its life cycle. The code to construct the preference page has been moved
out of TextureManipulator altogether, and now takes place during the
construction of MaterialManager, which makes the preference page visible
as soon as the application has started.