De-duplicate some code in the map format classes
MapFormat::canLoad now has a default implementation returning true, so
this does not need a definition in all but one of the map format
classes. The code to parse a map format version, common to both
Doom3MapFormat and Quake4MapFormat, is now refactored into a static
utility method Doom3MapFormat::hasMapVersion.