Disable initialisation of the Python module in unit tests
For an unknown reason, repeatedly initialising and finalising the Python
interpreter hundreds of times, as happens in unit tests, causes an
eventual std::exception and crash on Linux. This in turn causes the unit
test run to abort and fail, even though there is no actual problem with
whatever test happens to have triggered the crash.

Since none of the unit tests actually need Python functionality, this is
worked around by simply disabling initialisation of the ScriptingSystem
module in unit tests, via a new virtual method isPythonDisabled exposed
by the ApplicationContext.