ScriptingSystem is lazily initialised
Introduce the concept of a lazy module, exposed by a new virtual method
RegisterableModule::isLazy, with a default implementation that returns
false. Instead of being initialised at startup, lazy modules are
initialised when they are first requested by ModuleRegistry::getModule.

Currently the only lazy module is the ScriptingSystem, which is now
initialised when the GUI sets up the Script window and starts making
calls to GlobalScriptingSystem(). This means that the Python module is
not initialised at all during unit tests, replacing the earlier hack
which explicitly disabled Python via a (now removed) virtual method on
ApplicationContext.